finalizado

+++ Desafios/Ds07.py
@@ -1,4 +1,4 @@
-nota1 = int(input('Qual a primeira nota ?'))
-nota2 = int(input('Qual a segunda nota ?'))
+nota1 = float(input('Qual a primeira nota ?'))
+nota2 = float(input('Qual a segunda nota ?'))
 media = (nota1+nota2)/2
 print('A media das notas são {}'.format(media))
